#12 refactor: Merge unique landing page sections into single file
